+++ geotrek/api/tests/test_mobile/test_api_mobile.py
@@ -49,8 +49,8 @@ POI_LIST_PROPERTIES_GEOJSON_STRUCTURE = sorted([
 
 TOURISTIC_EVENT_LIST_PROPERTIES_GEOJSON_STRUCTURE = sorted([
     'id', 'name', 'description_teaser', 'description', 'themes', 'pictures', 'begin_date', 'end_date', 'duration',
-    'meeting_time', 'contact', 'email', 'website', 'organizer', 'speaker', 'type', 'accessibility', 'meeting_point',
-    'participant_number', 'booking', 'target_audience', 'practical_info', 'approved',
+    'start_time', 'contact', 'email', 'website', 'organizer', 'speaker', 'type', 'accessibility', 'meeting_point',
+    'capacity', 'booking', 'target_audience', 'practical_info', 'approved',
 ])
 
 TOURISTIC_CONTENT_LIST_PROPERTIES_GEOJSON_STRUCTURE = sorted([