+++ geotrek/api/v2/views/tourism.py
@@ -84,8 +84,9 @@ class TouristicEventViewSet(api_viewsets.GeotrekGeometricViewset):
     filter_backends = api_viewsets.GeotrekGeometricViewset.filter_backends + (
         api_filters.GeotrekTouristicEventFilter,
         api_filters.NearbyContentFilter,
-        api_filters.UpdateOrCreateDateFilter
+        api_filters.UpdateOrCreateDateFilter,
     )
+    filterset_class = api_filters.TouristicEventFilterSet
     serializer_class = api_serializers.TouristicEventSerializer
 
     def get_queryset(self):
@@ -98,3 +99,17 @@ class TouristicEventViewSet(api_viewsets.GeotrekGeometricViewset):
                               ) \
             .annotate(geom_transformed=Transform(F('geom'), settings.API_SRID)) \
             .order_by('name')  # Required for reliable pagination
+
+
+class TouristicEventPlaceViewSet(api_viewsets.GeotrekGeometricViewset):
+    filter_backends = api_viewsets.GeotrekGeometricViewset.filter_backends + (
+        api_filters.NearbyContentFilter,
+        api_filters.UpdateOrCreateDateFilter,
+        api_filters.TouristicEventsRelatedPortalFilter
+    )
+    serializer_class = api_serializers.TouristicEventPlaceSerializer
+
+    def get_queryset(self):
+        return tourism_models.TouristicEventPlace.objects.prefetch_related('touristicevents').annotate(
+            geom_transformed=Transform('geom', settings.API_SRID)
+        ).order_by('name')