+++ geotrek/tourism/views.py
@@ -3,7 +3,7 @@ import os
 
 from django.conf import settings
 from django.contrib.gis.db.models.functions import Transform
-from django.db.models import Q
+from django.db.models import Q, Sum
 from django.http import Http404
 from django.shortcuts import get_object_or_404
 from django.utils.translation import gettext as _
@@ -202,17 +202,22 @@ class TouristicEventFormatList(MapEntityFormat, TouristicEventList):
     mandatory_columns = ['id']
     default_extra_columns = [
         'structure', 'eid', 'name', 'type', 'description_teaser', 'description', 'themes',
-        'begin_date', 'end_date', 'duration', 'meeting_point', 'meeting_time',
-        'contact', 'email', 'website', 'organizer', 'speaker', 'accessibility',
-        'participant_number', 'booking', 'target_audience', 'practical_info',
+        'begin_date', 'end_date', 'duration', 'meeting_point', 'start_time', 'end_time',
+        'contact', 'email', 'website', 'organizer', 'speaker', 'accessibility', 'bookable',
+        'capacity', 'booking', 'target_audience', 'practical_info',
         'date_insert', 'date_update', 'source', 'portal',
         'review', 'published', 'publication_date',
         'cities', 'districts', 'areas', 'approved', 'uuid',
+        'cancelled', 'cancellation_reason', 'total_participants', 'place'
     ]
+
+    def get_queryset(self):
+        qs = super().get_queryset().select_related('place', 'cancellation_reason').prefetch_related('participants')
+        return qs.annotate(total_participants=Sum('participants__count'))
 
 
 class TouristicEventDetail(CompletenessMixin, MapEntityDetail):
-    queryset = TouristicEvent.objects.existing()
+    queryset = TouristicEvent.objects.existing().select_related('place', 'cancellation_reason').prefetch_related('participants')
 
     def get_context_data(self, *args, **kwargs):
         context = super().get_context_data(*args, **kwargs)
@@ -300,6 +305,8 @@ class TouristicEventViewSet(GeotrekMapentityViewSet):
         if self.format_kwarg == 'geojson':
             qs = qs.annotate(api_geom=Transform('geom', settings.API_SRID))
             qs = qs.only('id', 'name')
+        else:
+            qs = qs.select_related('place', 'cancellation_reason').prefetch_related('participants')
         return qs
 
     def get_columns(self):